Move the ID area to its own file
