+++ frontend/src/save-formats/N64/Components/Basics.js
@@ -1,0 +1,9 @@
+export default class N64Basics {
+  static LITTLE_ENDIAN = false;
+
+  static NUM_NOTES = 16; // A "note" is a save slot. It consists of >= 1 pages
+
+  static NUM_PAGES = 128;
+
+  static PAGE_SIZE = 256;
+}
